move retry_on_request_exceptions to utils and avoid Dataset import
