+++ tests/conftest.py
@@ -7,12 +7,12 @@ import tempfile
 import pytest
 from huggingface_hub import snapshot_download
 
-# from axolotl.utils.data.sft import retry_on_request_exceptions
-#
-#
-# @retry_on_request_exceptions(max_retries=3, delay=5)
-# def snapshot_download_w_retry(*args, **kwargs):
-#     return snapshot_download(*args, **kwargs)
+from axolotl.utils.data.utils import retry_on_request_exceptions
+
+
+@retry_on_request_exceptions(max_retries=3, delay=5)
+def snapshot_download_w_retry(*args, **kwargs):
+    return snapshot_download(*args, **kwargs)
 
 
 @pytest.fixture(scope="session", autouse=True)
